requirements.txt
aiofiles==23.2.1
aiohttp==3.9.1
aiosignal==1.3.1
aiosmtplib==3.0.1
annotated-types==0.6.0
anyio==3.7.1
async-fastapi-jwt-auth==0.6.1
async-timeout==4.0.3
asyncpg==0.29.0
attrs==23.1.0
certifi==2023.11.17
execnet==2.0.2
fastapi==0.100.1
frozenlist==1.4.1
greenlet==3.0.2
h11==0.14.0
httpcore==1.0.2
httpx==0.25.2
idna==3.6
iniconfig==2.0.0
multidict==6.0.4
packaging==23.2
passlib==1.7.4
Pillow==10.1.0
pluggy==1.3.0
pydantic==2.5.2
pydantic-settings==2.1.0
pydantic_core==2.14.5
PyJWT==2.8.0
pytest==7.4.3
pytest-xdist==3.5.0
python-dotenv==1.0.0
python-multipart==0.0.6
redis==5.0.1
sniffio==1.3.0
SQLAlchemy==2.0.23
SQLAlchemy-Utils==0.41.1
starlette==0.27.0
typing_extensions==4.9.0
yarl==1.9.4
